feat: made "getFullUsers" modular and reduced repeated code in components
