+++ public/index.html
@@ -2,14 +2,14 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/https://storage.googleapis.com/weppreciate-store/logo/weppreciate-logo-v1.png" />
+    <link rel="icon" href="https://storage.googleapis.com/weppreciate-store/logo/weppreciate-logo-v1.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
       name="We'ppreciate"
       content="You work hard. We'ppreciate you."
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/https://storage.googleapis.com/weppreciate-store/logo/weppreciate-logo-v1.png" />
+    <link rel="apple-touch-icon" href="https://storage.googleapis.com/weppreciate-store/logo/weppreciate-logo-v1.png" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
