+++ src/components/Dashboard/SendCard.jsx
@@ -1,7 +1,6 @@
 // Purpose: the logic and rendering for the modal popup for sending recognition
 
-import React, { useEffect, useState } from "react";
-import axios from "axios";
+import React, { useState } from "react";
 import {
   Alert,
   Button,
@@ -17,6 +16,7 @@ import { Send } from "@mui/icons-material";
 import { apiUrl } from "../../utils/ApiUrl";
 import teamValues from "../../utils/Values";
 import { jwtToken, userData } from "../../utils/LocalStorage";
+import FullUsers from "../../utils/FullUsers";
 
 export default function SendCard(props) {
   // Set default values of formData
@@ -28,36 +28,11 @@ export default function SendCard(props) {
   });
 
   // Establishing states
-  const [fullUsers, setFullUsers] = useState([]);
-  const [loading, setLoading] = useState(true);
   const [successMessage, setSuccessMessage] = useState("");
   const [errorMessage, setErrorMessage] = useState("");
 
-  // Importing full users list to render in form
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await axios.get(apiUrl + "users/all/fullusers", {
-          headers: {
-            Authorization: `Bearer ${jwtToken}`,
-          },
-        });
-
-        const sortedUsers = response.data.Users.sort((a, b) =>
-          `${a.name.first} ${a.name.last}`.localeCompare(
-            `${b.name.first} ${b.name.last}`
-          )
-        );
-
-        setFullUsers(sortedUsers);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchData();
-  }, []);
+  // Importing users data and loading state for nominations drop-down
+  const { fullUsers, loading } = FullUsers();
 
   const handleClick = () => {
     props.toggle();
@@ -67,7 +42,7 @@ export default function SendCard(props) {
   const handleSubmit = (event) => {
     event.preventDefault();
 
-    // establishing correct date format
+    // Establishing correct date format that's accepted by API
     const dateOptions = { day: "2-digit", month: "2-digit", year: "numeric" };
 
     // Takes card form data and converts into correct JSON format for POST request
@@ -158,107 +133,109 @@ export default function SendCard(props) {
       {errorMessage && <div className="errorMessage">{errorMessage}</div>}
       {successMessage && <div className="successMessage">{successMessage}</div>}
       <div className="modal">
-        {loading ? (
-          <div className="loader">
-            <CircularProgress />
-          </div>
-        ) : (
-          <div className="modal_content">
-            <span className="close" onClick={handleClick}>
-              &times;
-            </span>
-            <h2 className="formHeading">Let's send some recognition!</h2>
-            <form onSubmit={handleSubmit}>
-              <div className="cardForm">
-                <div className="formRow">
-                  <div className="formSelect">
-                    <TextField
-                      required
-                      name="recipient"
-                      className="formSelector"
-                      id="recipient"
-                      select
-                      label="Card recipient"
-                      defaultValue=""
-                      variant="outlined"
-                      onChange={handleChange}
-                    >
-                      {fullUsers.map((user) => {
-                        const userName = `${user.name.first} ${user.name.last}`;
-                        if (
-                          userName !==
-                          `${userData.name.first} ${userData.name.last}`
-                        ) {
-                          return (
-                            <MenuItem
-                              className="cardFormValues"
-                              key={user._id}
-                              value={user._id}
-                            >
-                              {userName}
-                            </MenuItem>
-                          );
-                        }
-                        // If there is a match with the logged in user's name, it won't be rendered on the list
-                        return null;
-                      })}
-                    </TextField>
-                  </div>
-                  <div className="formSelect">
-                    <TextField
-                      required
-                      name="value"
-                      className="formSelector"
-                      id="value"
-                      select
-                      label="Select a value"
-                      defaultValue=""
-                      variant="outlined"
-                      onChange={handleChange}
-                    >
-                      {teamValues.map((option) => (
-                        <MenuItem
-                          className="cardFormValues"
-                          key={option.value}
-                          value={option.value}
-                        >
-                          {option.label}
-                        </MenuItem>
-                      ))}
-                    </TextField>
-                  </div>
+        <div className="modal_content">
+          <span className="close" onClick={handleClick}>
+            &times;
+          </span>
+          <h2 className="formHeading">Let's send some recognition!</h2>
+          <form onSubmit={handleSubmit}>
+            <div className="cardForm">
+              <div className="formRow">
+                <div className="formSelect">
+                  <TextField
+                    required
+                    name="recipient"
+                    className="formSelector"
+                    id="recipient"
+                    select
+                    label="Card recipient"
+                    defaultValue=""
+                    variant="outlined"
+                    onChange={handleChange}
+                  >
+                    {loading ? (
+                      <div className="loader">
+                        <CircularProgress />
+                      </div>
+                    ) : (
+                      <div>
+                        {fullUsers.map((user) => {
+                          const userName = `${user.name.first} ${user.name.last}`;
+                          if (
+                            userName !==
+                            `${userData.name.first} ${userData.name.last}`
+                          ) {
+                            return (
+                              <MenuItem
+                                className="cardFormValues"
+                                key={user._id}
+                                value={user._id}
+                              >
+                                {userName}
+                              </MenuItem>
+                            );
+                          }
+                          // If there is a match with the logged in user's name, it won't be rendered on the list
+                          return null;
+                        })}
+                      </div>
+                    )}
+                  </TextField>
                 </div>
-                <TextField
-                  className="cardMessage"
-                  required
-                  multiline
-                  id="message"
-                  name="message"
-                  variant="outlined"
-                  label="Card message"
-                  value={formData.message}
-                  onChange={handleChange}
-                />
-                <FormControlLabel
-                  control={<Checkbox />}
-                  label="Nominate for award"
-                  id="award"
-                  name="award"
-                  checked={formData.award}
-                  onChange={handleChange}
-                />
-                <FormHelperText>
-                  Award nominations will be reviewed by managers
-                </FormHelperText>
+                <div className="formSelect">
+                  <TextField
+                    required
+                    name="value"
+                    className="formSelector"
+                    id="value"
+                    select
+                    label="Select a value"
+                    defaultValue=""
+                    variant="outlined"
+                    onChange={handleChange}
+                  >
+                    {teamValues.map((option) => (
+                      <MenuItem
+                        className="cardFormValues"
+                        key={option.value}
+                        value={option.value}
+                      >
+                        {option.label}
+                      </MenuItem>
+                    ))}
+                  </TextField>
+                </div>
               </div>
-              <div className="formButton">
-                <Button type="submit" variant="contained" endIcon={<Send />}>
-                  Submit
-                </Button>
-              </div>
-            </form>
-          </div>
-        )}
+              <TextField
+                className="cardMessage"
+                required
+                multiline
+                id="message"
+                name="message"
+                variant="outlined"
+                label="Card message"
+                value={formData.message}
+                onChange={handleChange}
+              />
+              <FormControlLabel
+                control={<Checkbox />}
+                label="Nominate for award"
+                id="award"
+                name="award"
+                checked={formData.award}
+                onChange={handleChange}
+              />
+              <FormHelperText>
+                Award nominations will be reviewed by managers
+              </FormHelperText>
+            </div>
+            <div className="formButton">
+              <Button type="submit" variant="contained" endIcon={<Send />}>
+                Submit
+              </Button>
+            </div>
+          </form>
+        </div>
       </div>
     </div>
   );
