+++ src/components/Settings/AddUser.jsx
@@ -8,11 +8,11 @@ import {
   MenuItem,
   TextField,
 } from "@mui/material";
-import { useEffect, useState } from "react";
-import axios from "axios";
+import { useState } from "react";
 
 import { apiUrl } from "../../utils/ApiUrl";
 import { jwtToken } from "../../utils/LocalStorage";
+import FullUsers from "../../utils/FullUsers";
 
 export default function AddUser(props) {
   // Set default values of formData
@@ -29,35 +29,11 @@ export default function AddUser(props) {
   });
 
   // Establishing states
-  const [fullUsers, setFullUsers] = useState([]);
-  const [loading, setLoading] = useState(true);
   const [successMessage, setSuccessMessage] = useState("");
   const [errorMessage, setErrorMessage] = useState("");
 
-  // Importing full users list to render in form
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await axios.get(apiUrl + "users/all/fullusers", {
-          headers: {
-            Authorization: `Bearer ${jwtToken}`,
-          },
-        });
-        const sortedUsers = response.data.Users.sort((a, b) =>
-          `${a.name.first} ${a.name.last}`.localeCompare(
-            `${b.name.first} ${b.name.last}`
-          )
-        );
-        setFullUsers(sortedUsers);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
+  // Importing users data
+  const { fullUsers, loading } = FullUsers();
 
   const handleClick = () => {
     props.toggle();
@@ -115,7 +91,6 @@ export default function AddUser(props) {
               </Alert>
             );
           }
-
           // Clear the error message after 5 seconds
           setTimeout(() => {
             setErrorMessage("");
@@ -176,147 +151,148 @@ export default function AddUser(props) {
       {errorMessage && <div className="errorMessage">{errorMessage}</div>}
       {successMessage && <div className="successMessage">{successMessage}</div>}
       <div className="modal">
-        {loading ? (
-          <div className="loader">
-            <CircularProgress />
-          </div>
-        ) : (
-          <div className="modal_content">
-            <span className="close" onClick={handleClick}>
-              &times;
-            </span>
-            <h2 className="formHeading">Add user</h2>
-            <form onSubmit={handleSubmit}>
-              <div className="cardForm">
-                <div className="formRow">
-                  <div className="formSelect">
-                    <TextField
-                      required
-                      id="firstName"
-                      name="firstName"
-                      variant="outlined"
-                      label="First name"
-                      className="formSelector"
-                      value={formData.firstName}
-                      onChange={handleChange}
-                    />
-                  </div>
-                  <div className="formSelect">
-                    <TextField
-                      required
-                      id="lastName"
-                      name="lastName"
-                      variant="outlined"
-                      label="Last name"
-                      className="formSelector"
-                      value={formData.lastName}
-                      onChange={handleChange}
-                    />
-                  </div>
-                </div>
-                <div className="formRow">
-                  <div className="formSelect">
-                    {/* add front-end validation on email regex */}
-                    <TextField
-                      required
-                      id="email"
-                      name="email"
-                      variant="outlined"
-                      label="Email"
-                      className="formSelector"
-                      value={formData.email}
-                      onChange={handleChange}
-                    />
-                  </div>
-                  <div className="formSelect">
-                    <TextField
-                      id="userPhotoKey"
-                      name="userPhotoKey"
-                      variant="outlined"
-                      label="Photo URL"
-                      className="formSelector"
-                      value={formData.userPhotoKey}
-                      onChange={handleChange}
-                    />
-                  </div>
-                </div>
-                <div className="formRow">
-                  <div className="formSelect">
-                    <TextField
-                      required
-                      id="businessUnit"
-                      name="businessUnit"
-                      variant="outlined"
-                      className="formSelector"
-                      label="Business unit"
-                      value={formData.businessUnit}
-                      onChange={handleChange}
-                    />
-                  </div>
-                  <div className="formSelect">
-                    <TextField
-                      name="userLineManager"
-                      className="formSelector"
-                      id="userLineManager"
-                      select
-                      label="Manager"
-                      defaultValue=""
-                      variant="outlined"
-                      onChange={handleChange}
-                    >
-                      {fullUsers.map((user) => {
-                        const userName = `${user.name.first} ${user.name.last}`;
-                        if (user.isLineManager) {
-                          return (
-                            <MenuItem
-                              className="cardFormValues"
-                              key={user._id}
-                              value={user._id}
-                            >
-                              {userName}
-                            </MenuItem>
-                          );
-                        }
-                        return null;
-                      })}
-                    </TextField>
-                  </div>
-                </div>
-                <div className="formRow">
-                  <FormControlLabel
-                    control={<Checkbox />}
-                    label="Admin"
-                    id="isAdmin"
-                    name="isAdmin"
-                    checked={formData.isAdmin}
+        <div className="modal_content">
+          <span className="close" onClick={handleClick}>
+            &times;
+          </span>
+          <h2 className="formHeading">Add user</h2>
+          <form onSubmit={handleSubmit}>
+            <div className="cardForm">
+              <div className="formRow">
+                <div className="formSelect">
+                  <TextField
+                    required
+                    id="firstName"
+                    name="firstName"
+                    variant="outlined"
+                    label="First name"
+                    className="formSelector"
+                    value={formData.firstName}
                     onChange={handleChange}
                   />
-                  <FormControlLabel
-                    control={<Checkbox />}
-                    label="Senior manager"
-                    id="isSeniorManager"
-                    name="isSeniorManager"
-                    checked={formData.isSeniorManager}
-                    onChange={handleChange}
-                  />
-                  <FormControlLabel
-                    control={<Checkbox />}
-                    label="Line manager"
-                    id="isLineManager"
-                    name="isLineManager"
-                    checked={formData.isLineManager}
+                </div>
+                <div className="formSelect">
+                  <TextField
+                    required
+                    id="lastName"
+                    name="lastName"
+                    variant="outlined"
+                    label="Last name"
+                    className="formSelector"
+                    value={formData.lastName}
                     onChange={handleChange}
                   />
                 </div>
               </div>
-              <div className="formButton">
-                <Button type="submit" variant="contained" endIcon={<Send />}>
-                  Add user
-                </Button>
+              <div className="formRow">
+                <div className="formSelect">
+                  <TextField
+                    required
+                    id="email"
+                    name="email"
+                    variant="outlined"
+                    label="Email"
+                    className="formSelector"
+                    value={formData.email}
+                    onChange={handleChange}
+                  />
+                </div>
+                <div className="formSelect">
+                  <TextField
+                    id="userPhotoKey"
+                    name="userPhotoKey"
+                    variant="outlined"
+                    label="Photo URL"
+                    className="formSelector"
+                    value={formData.userPhotoKey}
+                    onChange={handleChange}
+                  />
+                </div>
               </div>
-            </form>
-          </div>
-        )}
+              <div className="formRow">
+                <div className="formSelect">
+                  <TextField
+                    required
+                    id="businessUnit"
+                    name="businessUnit"
+                    variant="outlined"
+                    className="formSelector"
+                    label="Business unit"
+                    value={formData.businessUnit}
+                    onChange={handleChange}
+                  />
+                </div>
+                <div className="formSelect">
+                  <TextField
+                    name="userLineManager"
+                    className="formSelector"
+                    id="userLineManager"
+                    select
+                    label="Manager"
+                    defaultValue=""
+                    variant="outlined"
+                    onChange={handleChange}
+                  >
+                    {loading ? (
+                      <div className="loader">
+                        <CircularProgress />
+                      </div>
+                    ) : (
+                      <div>
+                        {fullUsers.map((user) => {
+                          const userName = `${user.name.first} ${user.name.last}`;
+                          if (user.isLineManager) {
+                            return (
+                              <MenuItem
+                                className="cardFormValues"
+                                key={user._id}
+                                value={user._id}
+                              >
+                                {userName}
+                              </MenuItem>
+                            );
+                          }
+                          return null;
+                        })}
+                      </div>
+                    )}
+                  </TextField>
+                </div>
+              </div>
+              <div className="formRow">
+                <FormControlLabel
+                  control={<Checkbox />}
+                  label="Admin"
+                  id="isAdmin"
+                  name="isAdmin"
+                  checked={formData.isAdmin}
+                  onChange={handleChange}
+                />
+                <FormControlLabel
+                  control={<Checkbox />}
+                  label="Senior manager"
+                  id="isSeniorManager"
+                  name="isSeniorManager"
+                  checked={formData.isSeniorManager}
+                  onChange={handleChange}
+                />
+                <FormControlLabel
+                  control={<Checkbox />}
+                  label="Line manager"
+                  id="isLineManager"
+                  name="isLineManager"
+                  checked={formData.isLineManager}
+                  onChange={handleChange}
+                />
+              </div>
+            </div>
+            <div className="formButton">
+              <Button type="submit" variant="contained" endIcon={<Send />}>
+                Add user
+              </Button>
+            </div>
+          </form>
+        </div>
       </div>
     </div>
   );
