+++ src/components/Settings/ManageUsers.jsx
@@ -12,53 +12,27 @@ import {
   CircularProgress,
 } from "@mui/material";
 
-import { useEffect, useState } from "react";
-import axios from "axios";
+import { useState } from "react";
 import { Settings } from "@mui/icons-material";
 
 import AddUserButton from "./AddUserButton";
-import { apiUrl } from "../../utils/ApiUrl";
 import EditUserButton from "./EditUserButton";
 import DeleteUserButton from "./DeleteUserButton";
-import { jwtToken, userData } from "../../utils/LocalStorage";
+import { userData } from "../../utils/LocalStorage";
 import ResetPasswordButton from "./ResetPasswordButton";
+import FullUsers from "../../utils/FullUsers";
 
 const ManageUsers = ({ setView }) => {
   const handleGoBackClick = () => {
     setView("default");
   };
 
-  // Establishing states
-  const [fullUsers, setFullUsers] = useState([]);
-  const [loading, setLoading] = useState(true);
+  // Establishing state
   // eslint-disable-next-line no-unused-vars
   const [selectedUser, setSelectedUser] = useState(null);
 
-  // Importing full users list to render in form
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await axios.get(apiUrl + "users/all/fullusers", {
-          headers: {
-            Authorization: `Bearer ${jwtToken}`,
-          },
-        });
-
-        const sortedUsers = response.data.Users.sort((a, b) =>
-          `${a.name.first} ${a.name.last}`.localeCompare(
-            `${b.name.first} ${b.name.last}`
-          )
-        );
-        setFullUsers(sortedUsers);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
+  // Importing users data
+  const { fullUsers, loading } = FullUsers();
 
   // Extracts full name of managers based on id
   function getFullName(userId) {
