+++ src/components/Settings/ReleaseAwards.jsx
@@ -19,6 +19,7 @@ import { Settings } from "@mui/icons-material";
 import { apiUrl } from "../../utils/ApiUrl";
 import { jwtToken } from "../../utils/LocalStorage";
 import ReleaseAwardButton from "./ReleaseAwardButton";
+import FullUsers from "../../utils/FullUsers";
 
 const ReleaseAwards = ({ setView }) => {
   const handleGoBackClick = () => {
@@ -28,7 +29,6 @@ const ReleaseAwards = ({ setView }) => {
   // Establishing states
   const [nominations, setNominations] = useState([]);
   const [loading, setLoading] = useState(true);
-  const [fullUsers, setFullUsers] = useState([]);
   // eslint-disable-next-line no-unused-vars
   const [selectedNom, setSelectedNom] = useState(null);
 
@@ -62,22 +62,8 @@ const ReleaseAwards = ({ setView }) => {
     fetchNominationData();
   }, []);
 
-  // Importing full users info
-  useEffect(() => {
-    const fetchFullUsers = async () => {
-      try {
-        const response = await axios.get(apiUrl + "users/all/fullusers", {
-          headers: {
-            Authorization: `Bearer ${jwtToken}`,
-          },
-        });
-        setFullUsers(response.data.Users);
-      } catch (error) {
-        console.error("Error fetching full users:", error);
-      }
-    };
-    fetchFullUsers();
-  }, []);
+  // Importing users data
+  const { fullUsers } = FullUsers();
 
   // Extracts full name for cards based on id for nominations
   function getFullName(userId) {
@@ -137,13 +123,17 @@ const ReleaseAwards = ({ setView }) => {
                           ? getFullName(nomination.nominatorFullUser)
                           : `${nomination.nominatorBasicUser.basicName.first} ${nomination.nominatorBasicUser.basicName.last}`}
                       </TableCell>
-                      <TableCell>{getFullName(nomination.recipientUser)}</TableCell>
+                      <TableCell>
+                        {getFullName(nomination.recipientUser)}
+                      </TableCell>
                       <TableCell>{nomination.nominationBody}</TableCell>
                       <TableCell>{nomination.nominationValue}</TableCell>
-                      <TableCell><ReleaseAwardButton
-                            nomination={nomination}
-                            onEdit={handleReleaseNomination}
-                          /></TableCell>
+                      <TableCell>
+                        <ReleaseAwardButton
+                          nomination={nomination}
+                          onEdit={handleReleaseNomination}
+                        />
+                      </TableCell>
                     </TableRow>
                   );
                 })}
